requirements.txt
PyYAML                        6.0.2
requests                      2.32.3
urllib3                       2.2.3
